Dimensions: moved to integration tests
